fix(opi-coding-agent): --config loading, auth exit code, inside_workspace, mutex safety
- H6: resolve_config now loads the --config file with precedence above
  project/user config but below individual CLI args.
- H9: Auth failures (missing API key) now exit with code 3 instead of 2.
- M5: ReadTool computes inside_workspace via canonicalize/starts_with.
- L4: Mutex::lock().unwrap() replaced with graceful error handling.
- Fix clippy: remove unnecessary u32 cast, collapse nested if-let.

+++ crates/opi-agent/src/lib.rs
@@ -290,7 +290,7 @@ pub async fn agent_loop(
         // H5: prepare_next_turn hook
         let next_turn_ctx = hooks::PrepareNextTurnContext {
             messages: messages.clone(),
-            turn: turn_idx as u32 + 1,
+            turn: turn_idx + 1,
         };
         if let Some(update) = hooks.prepare_next_turn(next_turn_ctx).await {
             messages.extend(update.extra_messages);
